requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==22.3.0
bleach==4.1.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
databases==0.4.3
dataframe-image==0.1.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.13
dnspython==2.2.1
docstring-parser==0.12
ecdsa==0.17.0
email-validator==1.1.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastapi==0.75.1
fastapi-pagination==0.9.3
filelock==3.4.2
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.3.0
gast==0.5.3
gdown==4.4.0
geographiclib==1.52
geopy==2.2.0
gino==1.0.1
gino-starlette==0.1.4
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.46.0
h11==0.13.0
h5py==3.6.0
html5lib==1.1
httptools==0.2.0
idna==3.3
imgkit==1.2.2
inflection==0.5.1
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==1.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaleido==0.2.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==14.0.1
lxml==4.7.1
Mako==1.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mongoengine==0.24.1
motor==2.5.1
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
orjson==3.6.7
orm==0.1.9
ormar==0.11.0
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
piccolo==0.34.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
plotly==5.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.1
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pymongo==3.12.3
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pypika-tortoise==0.1.5
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-pptx==0.6.21
pytz==2022.1
pywin32==303
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
seaborn==0.11.2
selenium==4.1.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.17.1
targ==0.3.7
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
tomli==2.0.1
tornado==6.1
tortoise-orm==0.19.1
tqdm==4.63.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typesystem==0.4.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.9
uvicorn==0.15.0
virtualenv==20.13.1
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.2
Werkzeug==2.1.2
widgetsnbextension==3.5.2
wrapt==1.14.1
wsproto==1.1.0
XlsxWriter==3.0.2
xlwt==1.3.0
yarl==1.7.2
